Added final txt
#Expert Diagnostic System
The expert diagnostic system uses the java expert shell system which is
a routine of commands used to create an intelligent system. Using the
CLIPS editor (and extension of JESS rules), it uses intelligent
behaviour to determine the problem with a typical business computer
program by allowing the user to determine what the issue is.

##JESS & Clips
http://clipsrules.sourceforge.net/index.html
http://www.jessrules.com/jess/